+++ src/components/Typography/_Paragraph/Paragraph.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import classNames from 'classnames';
+
+import './paragraph.scss';
+
+/**
+ * A library styled paragraph
+ */
+const Paragraph = ({ children, className }) => (
+  <p className={classNames('bi-p', className)}>{children}</p>
+);
+
+export default Paragraph;